chore(jetstream)!: split KeyValueBucket to interface, make config and status stateflows
